# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_domain_security.CreateAuthHostRequest:
    properties:
      description:
        type: string
      host:
        type: string
    type: object
  internal_domain_security.UpdateAuthHostConfigRequest:
    properties:
      enabled:
        type: boolean
      lockout_duration:
        type: integer
      max_attempts:
        type: integer
    type: object
  internal_domain_security.UpdateAuthHostRequest:
    properties:
      description:
        type: string
      host:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API 支持
  description: WAF 防火墙管理系统控制面 API 接口文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: WAF Control API
  version: "1.0"
paths:
...
  /security_management/auth_host_cfg:
    get:
      consumes:
      - application/json
      description: 获取当前授权主机全局配置
      produces:
      - application/json
      responses:
        "200":
          description: 授权主机配置
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 配置不存在
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取授权主机配置
      tags:
      - 安全管理
    put:
      consumes:
      - application/json
      description: 更新或创建授权主机全局配置
      parameters:
      - description: 授权主机配置信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_domain_security.UpdateAuthHostConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 更新授权主机配置
      tags:
      - 安全管理
  /security_management/auth_hosts:
    get:
      consumes:
      - application/json
      description: 查询所有授权主机信息
      produces:
      - application/json
      responses:
        "200":
          description: 授权主机列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取授权主机列表
      tags:
      - 安全管理
    post:
      consumes:
      - application/json
      description: 新增一个授权主机记录
      parameters:
      - description: 授权主机信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_domain_security.CreateAuthHostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 创建授权主机
      tags:
      - 安全管理
  /security_management/auth_hosts/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除授权主机
      parameters:
      - description: 授权主机ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 无效的ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 授权主机不存在
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 删除授权主机
      tags:
      - 安全管理
    get:
      consumes:
      - application/json
      description: 根据ID获取单个授权主机详情
      parameters:
      - description: 授权主机ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 授权主机详情
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 无效的ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 授权主机不存在
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取授权主机详情
      tags:
      - 安全管理
    put:
      consumes:
      - application/json
      description: 根据ID更新授权主机信息
      parameters:
      - description: 授权主机ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新的授权主机信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_domain_security.UpdateAuthHostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 更新授权主机
      tags:
      - 安全管理
...
  /system/query_pswd_status/{userId}:
    get:
      consumes:
      - application/json
      description: 根据用户ID查询密码状态信息
      parameters:
      - description: 用户ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 密码状态信息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 无效的用户ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 用户不存在
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取用户密码状态
      tags:
      - 安全管理
...
securityDefinitions:
  BearerAuth:
    description: 输入 Bearer {token} 格式的 JWT 令牌
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"